Add saturating artihmetic to SystemTime
This commit introduces three new methods to `std::time::SystemTime`:
* `SystemTime::saturating_add()`
* `SystemTime::saturating_sub()`
* `SystemTime::saturating_duration_since()`

It makes use of `SystemTime::MAX` and `SystemTime::MIN` which were
introcued in #148825, in order to avoid the reimplementation of a common
error handling pattern that happens in conjunction with these constants,
namely putthing the respective constant in an `Option::unwrap_or()` onto
the result of `SystemTime::checked_add()` and
`SystemTime::checked_sub()` respectively.

The `SystemTime::saturating_duration_since()` method may be a bit
special here, as it does not make use of these two constants but rather
saturates to `Duration::ZERO`, similar to the version found in
`std::time::Instant`.

Overall, this achives a greater consistency with the other modules found
in `std::time` as well as providing a nice way for handling artihmetic
errors while preventing interfacing code to reimplement the same logic
all over again.

Proposal ACP: <https://github.com/rust-lang/libs-team/issues/718>